Added support for actions above the form and above the list. Tha paginator also can be put before the list.
The warning flash has been added, also the notice, error and warning flashes have a '[notice|error|warning]_detail' which may be used to give a more detailed description of the message. The flashes are now easily translatable by using the '[notice|error|warning]_params' flash message which is passed as the second parameter to the '__()' function.



git-svn-id: http://svn.symfony-project.com/plugins/pmPropelGeneratorPlugin/trunk@31135 ee427ae8-e902-0410-961c-c3ed070cd9f9